Structure file naming for Custom Node installation
